requirements.txt
adios4dolfinx==0.9.4
fenics_dolfinx==0.9.0
fenics_ufl==2024.2.0
mpi4py==4.1.0
numpy==2.3.2
pandas==2.3.1
petsc4py==3.23.4
rich==14.1.0